ADD: 13 - Signout & Admin Dashboard

+++ admin-app/src/actions/auth.actions.js
@@ -1,9 +1,8 @@
 import { authConstants } from './constants';
 import axios from '../helpers/axios';
+import { useSelector } from 'react-redux';
 
 export const login = (user) => {
-  console.log(user);
-
   return async (dispatch) => {
     dispatch({ type: authConstants.LOGIN_REQUEST });
 
@@ -29,7 +28,7 @@ export const login = (user) => {
     } catch (err) {
       dispatch({
         type: authConstants.LOGIN_FAILURE,
-        payload: { error: 'server not responsing' },
+        payload: { error: 'server not responding' },
       });
     }
   };
@@ -45,19 +44,30 @@ export const isUserLoggedIn = () => {
         payload: { token, user },
       });
     } else {
-      dispatch({
-        type: authConstants.LOGIN_FAILURE,
-        payload: { error: 'Failed to login' },
-      });
+      console.log('Not logged in');
+      // dispatch({
+      //   type: authConstants.LOGIN_FAILURE,
+      //   payload: { error: 'Failed to login' },
+      // });
     }
   };
 };
 
 export const logout = () => {
   return async (dispatch) => {
-    localStorage.clear();
-    dispatch({
-      type: authConstants.LOGOUT_REQUEST,
-    });
+    dispatch({ type: authConstants.LOGOUT_REQUEST });
+    const res = await axios.post('/admin/signout');
+
+    if (res.status === 200) {
+      localStorage.clear();
+      dispatch({
+        type: authConstants.LOGOUT_SUCCESS,
+      });
+    } else {
+      dispatch({
+        type: authConstants.LOGOUT_FAILURE,
+        payload: { error: res.data.error },
+      });
+    }
   };
 };
